usecase added for random number trivia
First we created GetRandomNumberTrivia and then implemented it same as
the concrete one with one difference being that it does not accept any
arguments. So we created NoParams class in core usecase folder which
doesn't accept any argument. It was placed in core, because many
usecases may share condition where they don't accept any parameters